+++ src/backend/core/tests/test_models_items_root.py
@@ -64,6 +64,7 @@ def test_models_sub_item_abilities_downgraded():
         "media_auth": True,
         "download": True,
         "move": False,
+        "password_locked": False,
         "partial_update": True,
         "restore": False,
         "retrieve": True,
@@ -100,6 +101,7 @@ def test_models_sub_item_abilities_downgraded():
         "media_auth": True,
         "download": True,
         "move": False,
+        "password_locked": False,
         "partial_update": False,
         "restore": False,
         "retrieve": True,
@@ -135,7 +137,9 @@ def test_models_items_root_get_abilities_owner(
     item = factories.ItemFactory(
         users=[(user, "owner")], type=item_type, update_upload_state=upload_state
     )
-    link_select_options = LinkReachChoices.get_select_options(**item.ancestors_link_definition)
+    link_select_options = LinkReachChoices.get_select_options(
+        item.ancestors_link_reach, item.ancestors_link_role
+    )
     can_export = item_type == models.ItemTypeChoices.FOLDER
     expected_abilities = {
         "accesses_manage": True,
@@ -154,6 +158,7 @@ def test_models_items_root_get_abilities_owner(
         "media_auth": True,
         "download": True,
         "move": True,
+        "password_locked": False,
         "partial_update": True,
         "restore": True,
         "retrieve": True,
@@ -184,6 +189,7 @@ def test_models_items_root_get_abilities_owner(
         "media_auth": False,
         "download": False,
         "move": False,
+        "password_locked": False,
         "partial_update": False,
         "restore": True,
         "retrieve": True,
@@ -222,7 +228,9 @@ def test_models_items_root_get_abilities_administrator(
         filename=("document.pdf" if item_type == models.ItemTypeChoices.FILE else None),
         update_upload_state=upload_state,
     )
-    link_select_options = LinkReachChoices.get_select_options(**item.ancestors_link_definition)
+    link_select_options = LinkReachChoices.get_select_options(
+        item.ancestors_link_reach, item.ancestors_link_role
+    )
     can_export = item_type == models.ItemTypeChoices.FOLDER
     expected_abilities = {
         "accesses_manage": True,
@@ -241,6 +249,7 @@ def test_models_items_root_get_abilities_administrator(
         "media_auth": True,
         "download": True,
         "move": True,
+        "password_locked": False,
         "partial_update": True,
         "restore": False,
         "retrieve": True,
@@ -288,7 +297,9 @@ def test_models_items_root_get_abilities_editor_user(
         filename="document.pdf" if item_type == models.ItemTypeChoices.FILE else None,
         update_upload_state=upload_state,
     )
-    link_select_options = LinkReachChoices.get_select_options(**item.ancestors_link_definition)
+    link_select_options = LinkReachChoices.get_select_options(
+        item.ancestors_link_reach, item.ancestors_link_role
+    )
     can_export = item_type == models.ItemTypeChoices.FOLDER
     expected_abilities = {
         "accesses_manage": False,
@@ -307,6 +318,7 @@ def test_models_items_root_get_abilities_editor_user(
         "media_auth": True,
         "download": True,
         "move": False,
+        "password_locked": False,
         "partial_update": True,
         "restore": False,
         "retrieve": True,
@@ -341,7 +353,9 @@ def test_models_items_root_get_abilities_reader_user(
     user = factories.UserFactory()
     item = factories.ItemFactory(users=[(user, "reader")], type=item_type)
     access_from_link = item.link_reach != "restricted" and item.link_role == "editor"
-    link_select_options = LinkReachChoices.get_select_options(**item.ancestors_link_definition)
+    link_select_options = LinkReachChoices.get_select_options(
+        item.ancestors_link_reach, item.ancestors_link_role
+    )
     can_export = item_type == models.ItemTypeChoices.FOLDER
     expected_abilities = {
         "accesses_manage": False,
@@ -360,6 +374,7 @@ def test_models_items_root_get_abilities_reader_user(
         "media_auth": True,
         "download": True,
         "move": False,
+        "password_locked": False,
         "partial_update": access_from_link,
         "restore": False,
         "retrieve": True,